Refactor clientcontextfactory
