+++ src/backing/client/grpc_client_impl/BUILD
@@ -1,0 +1,41 @@
+#
+# Copyright 2020 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+load("@rules_cc//cc:defs.bzl", "cc_library", "cc_test")
+
+package(default_visibility = ["//src/backing/client:__pkg__"])
+
+cc_library(
+    name = "grpc_client_context_factory",
+    srcs = ["grpc_client_context_factory.cc"],
+    hdrs = ["grpc_client_context_factory.h"],
+    deps = [
+        "//src/backing/client:clock",
+        "//src/backing/client:grpc_client_options",
+        "@com_github_grpc_grpc//:grpc++",
+    ],
+)
+
+cc_test(
+    name = "grpc_client_context_factory_test",
+    size = "small",
+    srcs = ["grpc_client_context_factory_test.cc"],
+    deps = [
+        ":grpc_client_context_factory",
+        "//src/backing/client/testing_util:fake_clock",
+        "@com_google_googletest//:gtest_main",
+    ],
+)
